+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 2.2.1
+- Enhancement - Display custom fields in WooCommerce product PDF.
 - Fix - Fatal error when generating PDFs for posts with non-publish status.
 
 ## 2.2.0